refactored wheel(), merged nyan cat into music example
