+++ README.md
@@ -2,16 +2,15 @@ _The Internet Button has a lot of features, but some of these aren't as straight
 
 ## Examples
 
-- 1. __Blink an LED__ - How to control the smart LEDs on this board
-- 2. __Blink all the LEDs__ - An extension on the simplest case
-- 3. __LEDs and Buttons__ - How to read the buttons on the Button, and make LEDs blink with them
-- 4. __Good Combination__ - A set of button and LED conditionals that I happen to like and use frequently
-- 5. __Motion__ - The Internet Button also has an accelerometer on it to measure motion- this shows how the related functions work
-- 6. __Orientation__ - How to use the accelerometer functions to determine the orientation
-- 7. __Internet__ - Send that data out to the world!
-- 8. __Making Music__ - Learn how to play notes and songs with your Button
-- 8a. __Nyan Cat__ - Learn how to play a song while controlling LEDs and reading buttons at the same time.
-- 9. __Release Firmware__ - Big, complicated set of epic.
+1. __Blink an LED__ - How to control the smart LEDs on this board
+2. __Blink all the LEDs__ - An extension on the simplest case
+3. __LEDs and Buttons__ - How to read the buttons on the Button, and make LEDs blink with them
+4. __Good Combination__ - A set of button and LED conditionals that I happen to like and use frequently
+5. __Motion__ - The Internet Button also has an accelerometer on it to measure motion- this shows how the related functions work
+6. __Orientation__ - How to use the accelerometer functions to determine the orientation
+7. __Internet__ - Send that data out to the world!
+8. __Making Music__ - Learn how to play notes and songs while controlling LEDs and reading buttons at the same time.
+9. __Release Firmware__ - Big, complicated set of epic.
 
 
 _Watch for gauge and other display additions to this library!_
