+++ firmware/examples/8_MakingMusic.cpp
@@ -1,5 +1,113 @@
 // You know the routine by now, just going to comment on the new things!
 
+#include "InternetButton/InternetButton.h"
+
+InternetButton b = InternetButton();
+
+// This is our Nyan Cat song definition
+const char nyan_cat_intro[] = "DS5,16,E5,16,FS5,8,B5,8,DS5,16,E5,16,FS5,16,B5,16,CS6,16,DS6,16,CS6,16,AS5,16,B5,8,FS5,8,DS5,16,E5,16,FS5,8,B5,8,CS6,16,AS5,16,B5,16,CS6,16,E6,16,DS6,16,E6,16,B5,16\n";
+const char nyan_cat_melody[] = "FS5,8,GS5,8,DS5,16,DS5,16,R,16,B4,16,D5,16,CS5,16,B4,16,R,16,B4,8,CS5,8,D5,8,D5,16,CS5,16,B4,16,CS5,16,DS5,16,FS5,16,GS5,16,DS5,16,FS5,16,CS5,16,DS5,16,B4,16,CS5,16,B4,16,DS5,8,FS5,8,GS5,16,DS5,16,FS5,16,CS5,16,DS5,16,B4,16,D5,16,DS5,16,D5,16,CS5,16,B4,16,CS5,16,D5,8,B4,16,CS5,16,DS5,16,FS5,16,CS5,16,DS5,16,CS5,16,B4,16,CS5,8,B4,8,CS5,8,FS5,8,GS5,8,DS5,16,DS5,16,R,16,B4,16,D5,16,CS5,16,B4,16,R,16,B4,8,CS5,8,D5,8,D5,16,CS5,16,B4,16,CS5,16,DS5,16,FS5,16,GS5,16,DS5,16,FS5,16,CS5,16,DS5,16,B4,16,CS5,16,B4,16,DS5,8,FS5,8,GS5,16,DS5,16,FS5,16,CS5,16,DS5,16,B4,16,D5,16,DS5,16,D5,16,CS5,16,B4,16,CS5,16,D5,8,B4,16,CS5,16,DS5,16,FS5,16,CS5,16,DS5,16,CS5,16,B4,16,CS5,8,B4,8,CS5,8,B4,8,FS4,16,GS4,16,B4,8,FS4,16,GS4,16,B4,16,CS5,16,DS5,16,B4,16,E5,16,DS5,16,E5,16,FS5,16,B4,8,B4,8,FS4,16,GS4,16,B4,16,FS4,16,E5,16,DS5,16,CS5,16,B4,16,FS4,16,DS4,16,E4,16,FS4,16,B4,8,FS4,16,GS4,16,B4,8,FS4,16,GS4,16,B4,16,B4,16,CS5,16,DS5,16,B4,16,FS4,16,GS4,16,FS4,16,B4,8,B4,16,AS4,16,B4,16,FS4,16,GS4,16,E4,16,E5,16,DS5,16,E5,16,FS5,16,B4,8,AS4,8,B4,8,FS4,16,GS4,16,B4,8,FS4,16,GS4,16,B4,16,CS5,16,DS5,16,B4,16,E5,16,DS5,16,E5,16,FS5,16,B4,8,B4,8,FS4,16,GS4,16,B4,16,FS4,16,E5,16,DS5,16,CS5,16,B4,16,FS4,16,DS4,16,E4,16,FS4,16,B4,8,FS4,16,GS4,16,B4,8,FS4,16,GS4,16,B4,16,B4,16,CS5,16,DS5,16,B4,16,FS4,16,GS4,16,FS4,16,B4,8,B4,16,AS4,16,B4,16,FS4,16,GS4,16,B4,16,E5,16,DS5,16,E5,16,FS5,16,B4,8,CS5,8\n";
+
+// Prototypes of the functions that we'll use
+bool isPaused();
+void otherTasks();
+void playSongAndProcess(String song);
+
+void setup() {
+    b.begin();
+
+    // Set the BPM if you desire, the default is 250
+    b.setBPM(250);
+
+    // playSong takes a string in the format
+    // "NOTE,TYPE,NOTE,TYPE..."
+    // Types are note types that define duration so
+    // 16 is a 1/16th note, 8 is a 1/8th note and 4 is a 1/4 note, etc..
+    // (SEE A FULL DESCRIPTION OF playSong() at the end of this file)
+    // Here we will go up the complete scale in the 5th octave as 1/8th notes:
+    b.playSong("C5,16,C#5,16,D5,16,D#5,16,E5,16,F5,16,F#5,16,G5,16,G#5,16,A5,16,A#5,16,B5,16\n");
+
+    // playNote just plays a single note, and takes
+    // a note name as a string and a note type as an int
+    // Note types define duration like in scored music, so
+    // 16 is a 1/16th note, 8 is a 1/8th note and 4 is a 1/4 note, etc..
+    // And now we will go back down the complete scale using flat semi-tones
+    // instead of their sharp counterparts (5th octave as 1/8th notes):
+    b.playNote("Bb5",16);
+    b.playNote("A5",16);
+    b.playNote("Ab5",16);
+    b.playNote("G5",16);
+    b.playNote("Gb5",16);
+    b.playNote("F5",16);
+    b.playNote("E5",16);
+    b.playNote("Eb5",16);
+    b.playNote("D5",16);
+    b.playNote("Db5",16);
+    b.playNote("C5",16);
+    delay(500);
+
+    // Now let's get ready to play Nyan Cat
+    b.setBrightness(255); // 0 - 255
+    b.setBPM(142); // Nyan Cat BPM is 142
+    playSongAndProcess(nyan_cat_intro); // play the intro
+}
+
+void loop() {
+    // Now loop the melody forever! You may pause the song by pressing any button :)
+    playSongAndProcess(nyan_cat_melody);
+}
+
+
+
+/*********************************
+ * Below are our helper functions
+ *********************************/
+
+// If any button is pressed, toggle `paused` and return it's value
+bool isPaused() {
+    static bool paused = false;
+    if ( b.buttonOn(1) || b.buttonOn(2) || b.buttonOn(3) || b.buttonOn(4)) {
+        delay(50);
+        paused = !paused;
+        // wait until button is released
+        while ( b.buttonOn(1) || b.buttonOn(2) || b.buttonOn(3) || b.buttonOn(4));
+        delay(50);
+    }
+    return paused;
+}
+
+// Called once per note in the song
+void otherTasks() {
+    b.kickTheRainbow(25, 0);
+    while (isPaused()) {        // while paused...
+        b.setBrightness(50);    // dim the LEDs
+        b.kickTheRainbow(1, 2); // smoother rainbow
+        Particle.process();     // keep our Cloud connection going
+    }
+    b.setBrightness(255); // re-brighten the jam!
+}
+
+/* This is the InternetButton::playSong(String song) routine
+ * pulled out here so we can call kickTheRainbow() and read
+ * buttons in-between playing notes!
+ */
+void playSongAndProcess(String song){
+    char inputStr[song.length()];
+    song.toCharArray(inputStr,song.length());
+    char* note = strtok(inputStr,",");
+    char* duration = strtok(NULL,", \n");
+
+    while (duration != NULL) {
+        // Play the note, and then process otherTasks()
+        b.playNote(note, atoi(duration));
+        otherTasks();
+
+        // advance the note
+        note = strtok(NULL,",");
+        duration = strtok(NULL,", \n");
+    }
+}
+
 /*
  * b.playSong(String)
  *
@@ -21,51 +129,3 @@
  * b.playSong("C8,16,CS8,16,D8,16,DS8,16,E8,16,F8,16,FS8,16,G8,16,GS8,16,A8,16,AS8,16,B8,16\n");
  * b.playSong("C9,32,CS9,32,D9,32,DS9,32,E9,32,F9,32,FS9,32,G9,32,GS9,32,A9,32,AS9,32,B9,32,C9,32,CS9,32,D9,32,DS9,32,E9,32,F9,32,FS9,32,G9,32,GS9,32,A9,32,AS9,32,B9,32\n");
  */
-
-#include "InternetButton/InternetButton.h"
-
-InternetButton b = InternetButton();
-
-void setup() {
-    // uncomment to enable Serial logging to see note output
-    Serial.begin(9600);
-
-    b.begin();
-
-    // Set the BPM if you desire, the default is 250
-    b.setBPM(250);
-
-    // playSong takes a string in the format
-    // "NOTE,TYPE,NOTE,TYPE..."
-    // Types are note types that define duration so
-    // 8 is a 1/8th note and 4 is a 1/4 note
-    b.playSong("C4,8,E4,8,G4,8,C5,8,G5,4\n");
-}
-
-void loop() {
-    if(b.allButtonsOn()){
-        b.rainbow(10);
-        b.playSong("E5,8,G5,8,E6,8,C6,8,D6,8,G6,8\n");
-        b.allLedsOff();
-    }
-
-    if(b.buttonOn(1)){
-        // playNote just plays a single note, and takes
-        // a note name as a string and a note type as an int
-        // Note types define duration like in scored music, so
-        // 8 is a 1/8th note and 4 is a 1/4 note
-        b.playNote("G3",8);
-    }
-
-    if(b.buttonOn(2)){
-        b.playNote("G4",8);
-    }
-
-    if(b.buttonOn(3)){
-        b.playNote("G5",8);
-    }
-
-    if(b.buttonOn(4)){
-        b.playNote("G6",8);
-    }
-}
